Split up users controller spec
